fix: Handled bad threshold types on load
